requirements.txt
openai>=1.12.0
supabase>=2.3.0
python-dotenv>=1.0.0
pinecone>=3.0.0
langchain-openai>=0.0.5
numpy>=1.26.0 